feat: update to Go 1.24
- no longer using flyio
- update project structure
