+++ .github/workflows/build.yaml
@@ -1,0 +1,34 @@
+name: Build and Deploy Clair
+
+on:
+  push:
+    branches:
+      - master
+
+jobs:
+  release:
+    name: Build and Deploy Clair
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      packages: write
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v3
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Login to GitHub Container Registry
+        run: echo "${{ secrets.GITHUB_TOKEN }}" | docker login ghcr.io -u ${{ github.actor }} --password-stdin
+
+      - name: Build Docker image
+        shell: bash
+        run: |
+          CLAIR_IMAGE=ghcr.io/${{ github.repository }}:latest
+
+          docker build -t $CLAIR_IMAGE
+          docker push $CLAIR_IMAGE
