--- .github/workflows/fly.yml
@@ -1,23 +1,0 @@
-name: Deploy to Fly.io
-
-on:
-  push:
-    branches:
-      - master
-  repository_dispatch:
-
-env:
-  FLY_API_TOKEN: ${{ secrets.FLY_API_TOKEN }}
-
-jobs:
-  deploy:
-    runs-on: ubuntu-latest
-    steps:      
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Setuo flyctl
-        uses: superfly/flyctl-actions/setup-flyctl@master
-
-      - name: Deploy to Fly.io
-        run: flyctl deploy --remote-only
